Include real name for DM key updates (Closes: #693403)
We pull name/username details for DD/DN keys from the keyids file, but
don't have any username for a DM so don't have an equivalent file. We
could add one without usernames, but actually it just makes more sense to
pull the name from the key itself. If/when DMs get usernames we'll move
them over to using the keyids file as well.