Emit DAK commands to cleanup DM ACLs when a DM becomes a DD
ftp-master automatically import new keys into DAK when we add them to
the DM keyring, but the DM ACLs are not cleaned up when a key is moved
from the DM to DD keyring. Try and spot when we do this and emit the
cleanup commands in our DAK command file.