Update scripts to avoid exporting UATs (photo IDs)
As pointed out in #826713 the photo IDs included in some keys contain
copyrighted images. Rather than having to track exactly what license
each of these are under and selectively remove those that can't be
determined update the scripts to never output these packets - this
will help reduce the size of the keyring and the UAT packets are
unnecessary for the verification for which the keyring package is used.