Fix update-key to not duplicate keys
The changes to update-key that were intended to ensure our keyrings
stay in an internally verifiable state with regard to signatures (i.e.
with only signatures from keys that are contained within them)
resulted in duplicated keys, as both the existing and new key would
be exported. Alter this by importing into a copy of the keyring, then
exporting. This is slower, but has the side effect of giving more
useful import statistics as the signatures are cleaned on import, not
just export.