Add keyid to logline entry added by add-key
We've previously included the 64 bit key id in log entries for new keys.
While these are not cryptographically useful there's still some merit in
being able to easily see what key was added when from the changelog, so
add it back in to the log line that add-key now automatically generates.