Add script to parse email from Front Desk
This script attempts to parse emails from Front Desk requesting that
keys be added or moved between keyrings. It can be a bit fragile, so
it is important to manually eyeball what it has done, but it creates
a git-commit-template that tries to have as much detail pre-filled in
as possible.