Fix update-key to work with GnuPG 2
GnuPG 2 changes --with-colons to no longer put a UID on the pub line,
which breaks our method of obtaining names for DM who don't have a
username. Fix this, and also remove the setup of v1 keyrings as these
make GnuPG think it needs to do a migration from v1 -> v2.