Remove subkeys < 2048 bits and add a test to catch them
It was pointed out on #debian-devel that there were subkeys in the
keyring that were smaller than 2048 bits. Most of these are expired
or revoked, but 3 keys are affected. Remove these subkeys (manually
using gpg --edit-key / delkey) and extend our existing 2048 bit test
to cover subkeys as well as primary keys.