Make key size + expiry tests return failure when issues are found
The key size + expiry tests use awk to check if the error condition
was met. This was only printing a message when there was an issue,
rather than causing the test to return an error code that could be
picked up. Fix this so that auto tests during building catch these
keys properly. [Expired keys continue to just print a message and
not actually fail the tests.]