new make target for WKD: output/openpgpkey
I did not put this in the "all" target yet because that would have
required python and some modern version of gpg in the
build-dependencies, and i don't know how we want to deploy this.

I wanted to test to ensure that gniibe's key would be appropriately
small -- about 2KiB instead of the dozens of KiB that it was before i
laundered it.  And, it does the trick :)