scripts/update-keyrings: verify sha512sums against keyring-maint (and use gpgv)
The prior versions of this script just verified sha512sums.txt based
on any arbitrary OpenPGP certificate the user's keyring.

Instead, we want to be deliberate about whose certificates are
legitimate signers.  So we pull keyring-maint keys from the
system-available keyring, and use it as a curated keyring.

Using gpgv makes this process simpler, and bypasses many of gpg's
idiosyncrasies.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>