scripts/update-keyrings: ensure that the 5 known keyrings are checked
We want to make sure that the 5 known keyrings are actually checked.
If any of them aren't, that should be an error.  If any new files have
been included, then at least warn about it.

With this change, we can also remove the "keyrings/debian-keyring.gpg:
OK" output, since the human script operator no longer needs to inspect
for it.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>